requirements.txt
newspaper3k==0.2.8
PyQt5==5.15.4
textblob==0.15.3
